TFKUBE-43: updated unit test
